+++ api/src/main/java/org/openmrs/module/fhir2/api/translators/impl/BaseEncounterTranslator.java
@@ -9,14 +9,9 @@
  */
 package org.openmrs.module.fhir2.api.translators.impl;
 
-import java.util.Collections;
-import java.util.List;
-
 import lombok.AccessLevel;
 import lombok.Setter;
-import org.hl7.fhir.r4.model.CodeableConcept;
 import org.hl7.fhir.r4.model.Coding;
-import org.openmrs.EncounterType;
 import org.openmrs.Location;
 import org.openmrs.module.fhir2.FhirConstants;
 import org.openmrs.module.fhir2.api.mappings.EncounterClassMap;
@@ -41,29 +36,6 @@ public abstract class BaseEncounterTranslator {
 			coding.setCode(classCode);
 		}
 		return coding;
-	}
-	
-	protected List<CodeableConcept> mapEncounterTypeClass(EncounterType openmrsEncounterType) {
-		Coding coding = new Coding();
-		coding.setCode(openmrsEncounterType.getUuid());
-		coding.setDisplay(openmrsEncounterType.getName());
-		
-		CodeableConcept code = new CodeableConcept();
-		code.setCoding(Collections.singletonList(coding));
-		
-		return Collections.singletonList(code);
-	}
-	
-	protected EncounterType mapEncounterTypeField(CodeableConcept fhirEncounterType) {
-		EncounterType openmrsEncounterType = null;
-		
-		if (fhirEncounterType != null && fhirEncounterType.getCoding() != null && !fhirEncounterType.getCoding().isEmpty()) {
-			openmrsEncounterType = new EncounterType();
-			openmrsEncounterType.setName(fhirEncounterType.getCodingFirstRep().getDisplay());
-			openmrsEncounterType.setUuid(fhirEncounterType.getCodingFirstRep().getCode());
-		}
-		
-		return openmrsEncounterType;
 	}
 	
 }
